requirements.txt

# main
langchain>=0.0.123
openai>=0.27
llama-index>=0.4.20
html2text>=2020.1.16
streamlit>=1.20.0
dropbox>=11.36.0

# env variables
python-dotenv==1.0.0

# linting
black==22.12.0
isort==5.11.4
mypy==0.991
pylint==2.15.10